+++ app/opportunity-toolkit/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import OpportunityToolkitDashboard from "@/components/opportunity-toolkit/OpportunityToolkitDashboard";
+
+export default function OpportunityToolkitPage() {
+  return <OpportunityToolkitDashboard />;
+}